requirements.txt
tweepy
yfinance
python-dotenv
requests
beautifulsoup4
